Added: Program for Calculating Average Of Three Numbers

+++ basic-programs/CalulatingAverage.java
@@ -3,11 +3,11 @@ public class CalulatingAverage {
     public static void main(String[] args){
         Scanner sc = new Scanner (System.in);
         System.out.println("Enter Number 1: ");
-        double num1 = sc.nextInt();
+        double num1 = sc.nextDouble();
         System.out.println("Enter Number 2: ");
-        double num2 = sc.nextInt();
+        double num2 = sc.nextDouble();
         System.out.println("Enter Number 3: ");
-        double num3 = sc.nextInt();
+        double num3 = sc.nextDouble();
         double average = (num1+num2+num3)/3;
         System.out.println("The Average Is: "+average);
         sc.close();
